Set line width limit for imports to 100 (#1)
Problem: Haskell style guide states that all lines *should* be less
than 80 characters long. However, for imports this is not so practical,
because import lists are usually not really interesting and if the
limit is small, imports occupy more lines. It leads to more conflicts.
Solution: set line width limit for imports to 100.