[PBP-96] Assume that deriving strategies are old enough
Problem: we require deriving strategies for GHC-8.2.2 and newer.
This version is quite old and it's fine to assume that all new code
uses GHC-8.2.2 or newer. So it seems fine to always require them and
assume that people are more or less familiar with them.

Solution:
1. Rewrite that part of style guide to say that all deriving strategies
should be specified.
2. Update existing `deriving` usages accordingly.