[SRK-133] Drop the rule about LANGUAGE pragmas
Problem: after some discussion we concluded that the rule about
LANGUAGE pragmas (that they should be sorted and each one should
be on its line) is minor and is not worth being part of the style
guide. One reason is that we often enable a lot of default-extesions
and rarely specify pragmas in specific modules. Another reason is that
our style-guide is too big and it's hard to keep all of it in mind
anyway.

Solution: remove that rule.